+++ ScintSuite/SimulationCodes/Common/strikes.py
@@ -72,6 +72,9 @@ def readSINPAstrikes(filename: str, verbose: bool = False):
             'versionID1': np.fromfile(fid, 'int32', 1)[0],
             'versionID2': np.fromfile(fid, 'int32', 1)[0],
         }
+        logger.info('File %s'%filename)
+        logger.info('SINPA version: %i.%i'%(header['versionID1'],
+                                            header['versionID2']))
         if header['versionID1'] <= 4:
             # Keys of what we have in the file:
             header['runID'] = np.fromfile(fid, 'S50', 1)[:]
@@ -82,7 +85,7 @@ def readSINPAstrikes(filename: str, verbose: bool = False):
             header['FILDSIMmode'] = \
                 np.fromfile(fid, 'int32', 1)[0].astype(bool)
             header['ncolumns'] = np.fromfile(fid, 'int32', 1)[0]
-            if header['versionID1'] >= 4:
+            if header['versionID1'] >= 4 and plate.lower()!='collimator':
                 header['kindOfFile'] = np.fromfile(fid, 'int32', 1)[0]
             header['counters'] = \
                 np.zeros((header['nXI'], header['ngyr']), int)
@@ -117,12 +120,20 @@ def readSINPAstrikes(filename: str, verbose: bool = False):
                     if id_version < 0:
                         raise Exception('Not undestood SINPA version')
                 else:
-                    try:
-                        header['info'] = deepcopy(
-                            order[key_to_look][id_version][plate.lower()][header['kindOfFile']])
-                        found_header = True
-                    except KeyError:
-                        id_version -= 1
+                    if plate.lower() != 'collimator':
+                        try:
+                            header['info'] = deepcopy(
+                                order[key_to_look][id_version][plate.lower()][header['kindOfFile']])
+                            found_header = True
+                        except KeyError:
+                            id_version -= 1
+                    else:
+                        try:
+                            header['info'] = deepcopy(
+                                order[key_to_look][id_version][plate.lower()])
+                            found_header = True
+                        except KeyError:
+                            id_version -= 1
                     # if the id_version is already -1, just stop, something
                     # went wrong
                     if id_version < 0:
@@ -164,7 +175,7 @@ def readSINPAstrikes(filename: str, verbose: bool = False):
                 header['time'] = float(np.fromfile(fid, 'float32', 1)[0])
                 header['shot'] = int(np.fromfile(fid, 'int32', 1)[0])
         # Read the extra information included in version 2
-        if header['versionID1'] >= 2:
+        if header['versionID1'] >= 2 and plate.lower() != 'collimator':
             header['FoilElossModel'] = np.fromfile(fid, 'int32', 1)[0]
             if header['FoilElossModel'] == 1:
                 header['FoilElossParameters'] = np.fromfile(fid, 'float64', 2)
@@ -185,7 +196,7 @@ def readSINPAstrikes(filename: str, verbose: bool = False):
                 header['ScintillatorYieldParameters'] = \
                     np.fromfile(fid, 'float64', 2)
         # Calculate the radial position if it is a INPA simulation
-        if not header['FILDSIMmode']:
+        if not header['FILDSIMmode'] and plate.lower() != 'collimator':
             iix = header['info']['x0']['i']
             iiy = header['info']['y0']['i']
             for ig in range(header['ngyr']):
@@ -248,11 +259,8 @@ def readSINPAstrikes(filename: str, verbose: bool = False):
         except FileNotFoundError:
             header['geomID'] = None
             logger.warning('Not found SINPA namelist')
-        logger.info('File %s'%filename)
         logger.info('Total number of strike points: %i'%
                     np.sum(header['counters']))
-        logger.info('SINPA version: %i,%i'%(header['versionID1'],
-                                            header['versionID2']))
         logger.info('Average number of strike points per centroid: %i'%
                   int(header['counters'].mean()))
         return header, data